requirements.txt
Flask==2.0.1
viberbot==1.0.12
gunicorn==20.1.0
wit==6.0.0
requests==2.25.1
coverage==5.5
python-dotenv==0.18.0